+++ biotracker/src/TrackingThread.cpp
@@ -201,7 +201,7 @@ void TrackingThread::setFrameNumber(size_t frameNumber) {
     if (m_imageStream->setFrameNumber(frameNumber)) {
         MutexLocker trackerLock(m_trackerMutex);
         if (m_tracker) {
-            m_tracker->setCurrentFrameNumber(frameNumber);
+            m_tracker->setCurrentFrameNumber((int)frameNumber);
         }
         playOnce();
     }
@@ -211,7 +211,7 @@ void TrackingThread::nextFrame() {
     if (m_imageStream->nextFrame()) { // increments the frame number if possible
         MutexLocker trackerLock(m_trackerMutex);
         if (m_tracker) {
-            m_tracker->setCurrentFrameNumber(m_imageStream->currentFrameNumber());
+            m_tracker->setCurrentFrameNumber((int)m_imageStream->currentFrameNumber());
         }
     } else {
         m_playing = false;
@@ -353,7 +353,7 @@ void BioTracker::Core::TrackingThread::paint(const size_t w, const size_t h, QPa
 
     // clear background
     painter.setBrush(QColor(0, 0, 0));
-    painter.drawRect(QRect(0, 0, w, h));
+    painter.drawRect(QRect(0, 0, (int)w, (int)h));
     painter.setBrush(QColor(0, 0, 0, 0));
 
     m_paintMutex.lock();
@@ -386,7 +386,7 @@ void BioTracker::Core::TrackingThread::paint(const size_t w, const size_t h, QPa
         const float viewport_skew = ScreenHelper::calculate_viewport(
                                         m_texture.width(),
                                         m_texture.height(),
-                                        w, h, window, viewport
+                                        (int)w, (int)h, window, viewport
                                     );
 
         painter.setWindow(window);
